Use updated GitHub Actions environment file syntax
The nightly workflow dynamically sets the `RUSTFLAGS` environment
variable depending on host platform to enable symbol stripping in the
linker.

CVE-2020-15228 is a vulnerability in the GitHub Actions hosted runner
that will interpret arbitrary `set-env` and `add-path` sequences in
stdout logs, including untrusted output.

GitHub has stated that they are immediately deprecating these commands
and intend to remove them in the near future:

https://github.blog/changelog/2020-10-01-github-actions-deprecating-set-env-and-add-path-commands/

This PR replaces the `set-env` magic stdout sequence with an append to
the special `$GITHUB_ENV` file.