Use updated GitHub Actions path file syntax
The nightly workflow dynamically sets the `PATH` environment
variable on Windows to access a custom `win_flex_bison` artifact when
building mruby.

CVE-2020-15228 is a vulnerability in the GitHub Actions hosted runner
that will interpret arbitrary `set-env` and `add-path` sequences in
stdout logs, including untrusted output.

GitHub has stated that they are immediately deprecating these commands
and intend to remove them in the near future:

https://github.blog/changelog/2020-10-01-github-actions-deprecating-set-env-and-add-path-commands/

This PR replaces the `add-path` magic stdout sequence with an append to
the special `$GITHUB_PATH` file.