Update nightly builder to use `artichoke/ghaction-import-gpg@v4.1.0`
This change fixes some CVEs in the npm packages this action depends on.

Changes to the use of the action include:

- Specifiy the signing subkey fingerprint.
- Use the action step outputs to pass the key fingerprint to gpg.
- Rely on the signing key passphrase being loaded into the GPG agent by
  the action instead of echoing it to gpg over stdin.

Further simplifications of gpg usage come from moving gpg signing out of
the build step by relying on build step asset outputs to perform the
signing in one place.

Further work on #32. Once this is merged and the nightly builder runs
successfully, #32 can be closed out and completed.