Upgrade `ghaction-import-gpg` to v5.1.0
`artichoke/ghaction-import-gpg` is synced from upstream which includes
several new releases, including a breaking change to v5 which upgrades
the nodejs runtime version.

The tag in `artichoke/ghaction-import-gpg` is:

- https://github.com/artichoke/ghaction-import-gpg/tree/v5.1.0

Diff of the upgrade from v4.1.0 to v5.1.0 is:

- https://github.com/artichoke/ghaction-import-gpg/compare/v4.1.0..v5.1.0.

The diff is mostly tests and linting, with some changes to how key
fingerprints and subkeys are handled, most notably:

- https://github.com/crazy-max/ghaction-import-gpg/pull/123
- https://github.com/crazy-max/ghaction-import-gpg/pull/125
- https://github.com/crazy-max/ghaction-import-gpg/pull/129

`artichoke/nightly` uses a signing subkey, so these fixes are good to
take.