Reduce maintenance burden of `ghaction-import-gpg`
`ghaction-import-gpg` is a privileged GitHub Action that is passed key
material for codesigning. To protect against supply chain risk, the
nightly workflow used a fork of upstream pinned to a recent-ish tag. The
goal with using a fork was to protect against tag mutability.

GitHub Actions supports pinning by Git SHA and _also_ supports updating
a comment that translates that SHA to a tag via dependabot.

This commit updates the CI and nightly workflows to depend on upstream
as of the same tag (v5.2.0, the latest release) but pinned by SHA.
Subsequent version bumps will be handled by dependabot with the
auditability and informatative PRs that come with that.

After this is merged, I plan to delete the forked repository at
https://github.com/artichoke/ghaction-import-gpg